Refactor state handlers
